+++ NEWS.md
@@ -1,3 +1,11 @@
+## here 0.0-6 (2017-01-16)
+
+- New `set_here()` function that creates a `.here` file and talks about it by default (#1).
+- New `dr_here()`, describes why `here()` has settled for a particular path.
+- Recognize projectile projects and VCS roots.
+- Using working directory as fallback produces wrong results, reverted.
+
+
 ## here 0.0-5 (2016-10-29)
 
 - `remake` projects are also recognized by default.